Fix multi-node
This mainly introduces the `num_nodes` argument and fixes all
occurrences of `TrainArgs` method calls that were previously only
supplied the number of devices instead of the world size (which we now
access using `fabric.world_size`).

Because we no longer need to pass the `devices` integer around, this
argument is removed from all signatures.